Split vuex modules into smaller files
